Added Create Encrypted File button
Create File opens a new window and asks for a file name and encryption
password, then creates a new text file of the name given and opens it.
When the user is finished, they save and close the file. At which point,
it's encrypted using the password they gave. All options for normal
encryption still apply.